allow addnl scopes in github & gitlab connectors
Allow additional scopes to be specified in GitHub and GitLab connectors that the connector just passes through during the OAuth flow. The idea is to use Dex as the single point of authentication for an application that has tight integration with GitHub and GitLab. Not need to replicate the authentication flow in the application. The resulting access token will have those permissions, should the user accept them, and they can be examined and used by a third party application.