optional additional scopes for GitHub & GitLab
This would allow a list of additional scopes to be specified in connector configuration of github and gitlab connectors. The scopes listed here will be passed additionally in the GitHub/GitLab OAuth2 flow. E.g.:

```
- type: github
  id: github
  name: GitHub
  config:
    clientID:
    clientSecret:
    redirectURI:
    additionalScopes:
    - notifications
    - read:public_key
```

These are normally not required. But useful for applications that are more tightly integrated with Dex and would want to reuse the same access token that Dex obtained for additional purposes. This would allow for one seamless signin flow.